requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
mistralai==0.0.12
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.1